requirements.txt
fastapi
uvicorn
pymongo
python-jose
pydantic
click
python-dotenv
python-multipart